add validation to delete the user

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@
 namespace App\Http\Controllers;
 
 use App\User;
+use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
@@ -13,8 +14,11 @@ class UserController extends Controller
      * @param int $id
      * @return  \Illuminate\Http\Response
      */
-    public function delete(User $user)
+    public function delete(User $user, Request $request)
     {
+        if($request->user()->id === $user->id) {
+            return redirect()->route('home');
+        }
         $user->delete();
         return redirect()->route('home');
